+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@ _TBD_
 * Integrated CQM c3/c4 access
   * Updated reports to display changes
   * Display / allow edit of c3/c4 for 2015 product CQMS
+* Updated links in /overview
 
 ---